requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pytesseract
Pillow
google-genai
python-dotenv
langchain
langchain-google-genai
langchain-openai
langchain-huggingface
pdf2image
langchain-community
faiss-cpu
tiktoken
langchain-text-splitters
